Token config helper function moved into scripts/tokens directory
