+++ src/screens/main/tokens/add/Tokens.js
@@ -9,7 +9,7 @@ import { NavigationActions } from 'react-navigation';
 import RF from 'react-native-responsive-fontsize';
 import * as actions from '../../../../actions/ActionCreator';
 import LinearButton from '../../../../components/LinearGradient/LinearButton';
-import TokenConfig from '../../../../scripts/tokenConfig';
+import TokenConfig from '../../../../scripts/tokens/tokenConfig';
 import { addNewToken } from '../../../../actions/AppConfig';
 
 /**
