+++ src/screens/setup/terms/TermsAndConditions.js
@@ -8,7 +8,7 @@ import { NavigationActions } from 'react-navigation';
 import { initializeAppTokenState, enterDebug } from '../../../actions/AppConfig';
 import { Terms } from '../../../constants/Terms';
 import LinearButton from '../../../components/LinearGradient/LinearButton';
-import TokenConfig from '../../../scripts/tokenConfig';
+import TokenConfig from '../../../scripts/tokens/tokenConfig';
 
 class TermsAndConditions extends Component {
   /**
